restructure file tree
- tried an experiemental file tree structure that I didn't like
- after being annoyed for a while - moved back to a more redundant naming convention
