--- src/app.container/style.scss
@@ -1,100 +1,0 @@
-@import '../theme/reset';
-@import '../theme/icons';
-@import '../theme/typography';
-@import '../theme/swatches';
-
-$transition-sky: 500ms;
-$transition-sun-moon: 300ms;
-$size: 200;
-$scaled-small: $size * ($small * 0.01);
-$scaled-very-small: $size * ($very-small * 0.01);
-$scaled-medium: $size * ($medium * 0.01);
-$scaled-medium-large: $size * ($medium-large * 0.01);
-$scaled-large: $size * ($large * 0.01);
-
-.container {
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-  justify-content: space-between;
-}
-
-// On mount the sun and moon spin crazy fast from 0 to the current rotation
-// to prevent this we wait untill the component has mounted and put everything in place
-// before initalizing transform.
-.hasMounted {
-  .sunMoon,
-  .sun,
-  .moon {
-    transition: transform $transition-sun-moon linear;
-  }
-}
-
-.timeKeeper {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 24px 0;
-}
-
-.timers {
-  padding: 24px;
-  max-width: 320px;
-  margin: 0 auto;
-  border-top: 1px solid $swatch01;
-}
-
-.controls {
-  display: flex;
-  width: $size * 1.38px;
-  justify-content: space-between;
-  margin-top: 32px;
-  @include font((font-size: $scaled-very-small, font-weight: $normal));
-
-  .button {
-    color: $swatch01;
-    transition: color 300ms ease;
-
-    &:disabled {
-      color: $swatch04;
-
-      &:hover {
-        color: $swatch04;
-      }
-
-      &:active {
-        color: $swatch04;
-      }
-    }
-
-    &:hover {
-      color: $swatch05;
-    }
-
-    &:active {
-      color: $swatch07;
-    }
-  }
-}
-
-.dayCount {
-  text-align: center;
-  @include font((font-size: $scaled-small, font-weight: $bold));
-  margin-bottom: $size * 0.1px;
-}
-
-.time {
-  @include font((font-size: $scaled-medium-large));
-  margin-bottom: $size * 0.1px;
-}
-
-.colon {
-  margin: 0 $size * 0.1px;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: row;
-  position: relative;
-}
