+++ src/components/increment-button/increment-button.component.style.scss
@@ -1,0 +1,51 @@
+@import '../../theme/typography';
+@import '../../theme/swatches';
+@import '../../theme/scale';
+
+.container {
+  margin-right: 8px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+
+  &:last-child {
+    margin-right: 0;
+  }
+}
+
+.button {
+  cursor: pointer;
+  width: 40px;
+  height: 40px;
+  border-radius: 100%;
+  background-color: $swatch01;
+  color: $swatch02;
+  @include font((font-size: $medium));
+  margin-right: 8px;
+  transition: background-color 300ms ease;
+
+  &:last-child {
+    margin-right: 0;
+  }
+
+  &:focus {
+    outline: 0;
+  }
+
+  &:hover {
+    background-color: $swatch05;
+  }
+
+  &:active {
+    background-color: $swatch07;
+  }
+}
+
+.duration {
+  font-weight: $bold;
+}
+
+.unit {
+  font-weight: $bold;
+}
