+++ src/components/sundial/sundial.component.style.scss
@@ -1,0 +1,156 @@
+@import '../../theme/icons';
+@import '../../theme/typography';
+@import '../../theme/swatches';
+
+$transition-sky: 500ms;
+$transition-sun-moon: 300ms;
+$size: 200;
+
+.sundial {
+  height: $size * 0.74px;
+  width: $size * 1.38px;
+  overflow: hidden;
+  margin-bottom: $size * 0.05px;
+  background-image: url('../../assets/images/sundial-frame.svg');
+  background-repeat: no-repeat;
+  position: relative;
+}
+
+.sky {
+  height: $size * 0.65px;
+  width: $size * 1.28px;
+  border-radius: $size * 1.28px $size * 1.28px 0 0;
+  position: absolute;
+  z-index: -10;
+  top: $size * 0.04px;
+  left: $size * 0.05px;
+  overflow: hidden;
+  transition: background $transition-sky linear;
+
+  @mixin dusk-till-dawn {
+    content: '';
+    height: $size * 0.64px;
+    width: $size * 1.28px;
+    border-radius: $size * 1.28px $size * 1.28px 0 0;
+    position: absolute;
+    top: 0;
+    left: 0;
+    overflow: hidden;
+    opacity: 0;
+    transition: opacity $transition-sky linear;
+  }
+
+  // dusk
+  &::before {
+    @include dusk-till-dawn;
+    background: linear-gradient(to bottom, $swatch21 0%, $swatch10 100%);
+  }
+
+  // dawn
+  &::after {
+    @include dusk-till-dawn;
+    z-index: -1;
+    background: linear-gradient(to bottom, $swatch21 0%, $swatch19 100%);
+  }
+
+  &.night {
+    background: $swatch20;
+  }
+
+  &.dusk {
+    &::before {
+      opacity: 1;
+    }
+  }
+
+  &.dawn {
+    &::after {
+      opacity: 1;
+    }
+  }
+
+  &.day {
+    background: $swatch21;
+
+    .stars {
+      opacity: 0;
+    }
+  }
+}
+
+.stars {
+  height: $size * 0.64px;
+  width: $size * 1.28px;
+  background-image: url('../../assets/images/sundial-stars.svg');
+  background-repeat: no-repeat;
+  position: absolute;
+  top: $size * 0.07px;
+  left: $size * 0.09px;
+}
+
+.sunMoon {
+  width: $size * 1.1px;
+  height: $size * 1.1px;
+  position: absolute;
+  left: $size * 0.09px;
+  top: $size * 0.09px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  transition: transform $transition-sun-moon linear;
+}
+
+.sun {
+  color: $swatch22;
+  filter: blur(2px);
+  @include fontsmoothing;
+
+  &::before {
+    @include icon('moon15', 'moon');
+    margin-right: 0;
+    font-size: $size * 0.28px;
+  }
+}
+
+.moon {
+  color: $swatch23;
+  @include fontsmoothing;
+  position: relative;
+  transition: transform $transition-sun-moon linear;
+
+  &::before {
+    margin-right: 0;
+    font-size: $size * 0.28px;
+    position: relative;
+    left: 25%;
+    z-index: 1;
+  }
+
+  &::after {
+    @include icon('moon15', 'moon');
+    margin-right: 0;
+    font-size: $size * 0.28px;
+    position: relative;
+    left: -25%;
+    color: transparent;
+    transition: color 0 linear;
+  }
+}
+
+@for $i from 2 through 28 {
+  .moon#{$i} {
+    &::before {
+      @include icon('moon#{$i}', 'moon');
+    }
+  }
+}
+
+.night {
+  .moon {
+    &::after {
+      color: $swatch20;
+      transition-delay: $transition-sky;
+    }
+  }
+}
