+++ src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import timeReducer from './time.reducer';
-import timersReducer from './timers.reducer';
+import timeReducer from './time/time.reducer';
+import timersReducer from './timers/timers.reducer';
 
 const rootReducer = combineReducers({
   timeState: timeReducer,
